Split import errors into a separate table
Using a serialized column for import errors was error prone and had bad
performance. Each time an error was logged it would need to deserialize
the column, append the error, reserialize and then re-transmit the whole
field to the database server. On a few occasions the serialized data has
overflowed the 64kb limit of a default text field.