Don't serve inline attachments from assets host
The asset host is primarily for inline images. By serving attachments
from the asset host, we're preventing authentication cookies from being
passed through for the attachment URLs. That means that the links don't
work when previewing a draft document.

I've changed the govspeak helper so that inline attachments use a
relative URL instead, and will therefore be served from the same domain
as the page itself.