Fix tests looking for roles rather than people
Using the people/person partial had changed how the roles were presented, this fixes the tests to look for the current_person and not the role.

Also added a conditional around the partial to skip roles that do not have a person assigned. We will look into the need case for this and will probably change this so that unassigned roles display with a notice.