Add chief scientific advisor role type
This isn't quite mirroring the special respresentatives and traffic
commissioners from previous changes, as the scientific advisors are
being shown in the management team at the moment. I've renamed the
`board_member_` variables to `management` to better reflect the grouping
that it represents.

Also taken the opportunity to remove some has_many statements which are
now redundant through the addition of ordering to the admin interface.