bumping Plek to 1.0.0
In doing this I detected some tests that are assuming GOVUK_ASSET_HOST
is not set (or set empty). We're in the process of trying to establish a
standard way of setting up environment variables for each app (currently
govuk_setenv is available) therefore I have modified the tests to
function correctly when this variable is set. When not set,
NoConfiguration exception will be thrown - same goes for
GOVUK_APP_DOMAIN.

@heathd worked with me on some of this