Script to move no-extension attachments to S3
This script transforms an incoming data file to by moving any
attachments without an extension to an S3 location with a guessed
extension, and rewrites the columns to use the new location.

It has the side effect of putting bad s3 urls in the redirector, but
that was considered acceptable.

Committed here for posterity and future reference.