Reset all NewsArticle and FatalityNotice documents
There are some erroneously versioned announcements on production that
need to be rolled up to be minor changes. This migration finds all
documents that match those criteria and resets all their editions to the
same minor version. This has the effect of removing the change notes
from the frontend and resetting the display/sort date to a sensible and
useful value.