CSV of GOV.UK exempt websites
We are publishing a CSV containing the list of
websites that are exempt from GOV.UK (list 
accurate as of 2012-12-17).

This file is not intended to be publicly exposed
on GOV.UK but will be linked to from the GDS blog
post announcing the publication of the list.

This is a temporary location for this file
(agreed with @jystewart). It was felt that the
whitehall repo was the most appropriate place for
it.