Extract iteration of documents to generate feed.entry calls to FeedHelper.
It makes sense to standardise how we iterate over documents and generate
atom feed.entry elements of those documents.  This way, the atom.builder
files for generating feeds are as similar as possible.  We also
standardise on using timestamp_for_sorting as the published timestamp and
timestamp_for_update as the updated timestamp on the entry elements.