FeedHelper#documents_as_feed_entries now sets the feed updated element.
All the feeds were doing the same thing; check if there are any docs, if
so use the updated timestamp of the first document, if not set it to some
other useful value.  Took the opportunity to extract this and put it in
one place because there were a couple of minor discrepancies (Time.now vs.
Time.zone.now, some not setting a default if no documents, etc...).