Switch font mixins from `core` to `ig-core`
This is the first step in switching to the new frontend toolkit fonts.
In the toolkit there are a different set of fonts mixins which use the
core name. This will let us continue to use our cores while we migrate
each page to the new stuff. Eventually we should be left with no
references to `ig-core` and at that point we can remove them entirely.