Editions can go from imported->draft state
They also have methods to ask if that is possible, and it should only be
true for imported models that would be valid as a draft. We changed
the validations on Publications so that imported publications are not
valid if they haven't had their publication_type set to something other
than ImportedAwaitingType.