Show temporary organisation if no deliverer yet
If the deliverer is not yet set, we want to show the organisation that
has been temporarily assigned during the import process. This is so that
we can reassign the edition if necessary to the right team for
processing, and to ensure that the organisation is preserved across
saves if the deliverer is not immediately set.