Hide joining bar with count in JavaScript not Ruby
Due to the caching that we have in production the bar would always
be hidden if the first person that visits the page in a cache window had
already dissmissed the bar.

Moved the logic into the frontend so it should never be effected by
the cache. Also used the `.js-hidden` class so that you will never see
the flash of content if you have already dismissed the bar and it will
instead just get shown to people who have not dismissed the bar.