Redirect admin/editions/X to correct controller
After getting tired of having to look up the type of an edition to be
able to enter the URL for it, I made this change to redirect all the
edition types from the main edition controller to their sub controllers.

E.g. for a publication with id 12345, /admin/editions/12345 will now
redirect to /admin/publications/12345 rather than 404ing.