Remove first_published_at for some types
This should remove the data in `first_published_at` for consultations,
publications and speeches. It is currently populated with
`major_change_published_at` for the first major edition of the
respective document.

Having data in the column is confusing as we won't be putting data in
for new editions. For those edition types the `first_public_at` method
pulls data from different columns.