Data migration for lead organisations on editions
We make the first organisation on an edition without any lead
organisations already into the first lead organisation for the edition.
We output the results as a CSV that includes the URL for any editions that
need more attention (e.g. have errors, have more than one org).  Can be
run in VERBOSE mode to get output for *all* the editions.