Remove duplication of policies in the speed tagging UI
The view code still leaves open the possibility of duplication in the UI of role appointments or ministerial roles.  I'm not sure, but I think that's a presentational thing where we show someone's name twice, but it's because of them having multiple roles.  Would like to see someone report an issue before fixing it.