Allow swapping a organisation from supporting to lead (and vice-versa)
Surprisingly, our hand-rolled uniqueness check for edition organisations had an edge-case.  The fix is that when collecting all the edition_organisation instances for comparison purposes we need to remove any instances that are destroyed? or marked_for_destruction?.  Crucially we have to remove them at the point of collection and also remove any duplicate instances from the other 2 associations.  This is because we're likely to only mark the instance for destruction on 1 of the associations, but it'll be present in another one.  E.g. an edition_organisation instance representing a lead organisation is present in the edition_organisations association and has a clone in the lead_edition_organisations association).