EditionControllers that allow attachments cope with attachment_action
They will massage the provided params to make sure that:

* 'keep' attachments aren't inadvertently removed or changed by removing the attachment_data_attributes and _destroy params
* 'remove' attachments are destroyed by adding the _destroy param
* 'replace' attachments are definitely creating a new attachment_data by removing any id param inside attachment_data_attributes

We don't do anything if the edition_attachments_attributes params don't include an id param as that's creating a brand new attachment and these keys shouldn't be present.

Note: we have place-holders for AttachmentData#to_replace_id= and AttachmentData#replaced_by so for now, the tests still fail.