Bugfix: avoid wrong link on announcments index
When a topic and organisation filter were combined the wrong document
would get eager-loaded, and therefore the wrong hyperlink would be used
on the announcements index page.

The bug was introduced during the work to add topical_events. The
problem was that the .published_in_topic class method introduced by the
Edition::Classifications mixin was clobbering the .published_in_topic
method of Edition::RelatedPolicies. The .published_in_topic method
should never have been defined on Edition::Classifications, it belongs
on Edition::Topics.