AttachmentData instances can now be replaced
AttachmentData#replace_with! will set the (new) replaced_by association on the instance.  It also walks the object graph backwards so that replacing an instance (x) that is already a replacement will set the instances that x replaced to be replaced_by the instance we are replacing x with.

We can trigger all this from the original instance by setting the virtual attribute to_replace_id.  During save this loads the AttachmentData instance referenced and calls replace_with! on it.