World locations have contacts
Extracted org tests from model and controller and applied them to the
worldwide office object. Tweaked the contact model to allow polymorphic
associations.

Note that the migration will copy the organisation information into the
new contactable fields but will leave the organisation_id field as an
orphan to be removed in a future migration.