Features tagged '@quarantine-files' uploads behave more like production
We stub Whitehall.clean_upload_path to return a new, test-specific, location and arrange Carrierwave.incoming_root and Carrierwave.clean_root and the virus-checking step to be more production like.  E.g. files that have been virus checked are not placed into public/.  This means that rails default ActionDispatch::Static middleware doesn't jump on incoming urls for attachments and avoid the AttachmentsController and so we can test the real behaviour.