Update tests using major_change_published_at
Lots of tests were still using major_change_pubished_at in an attempt to
sort editions. I have updated those test. In many cases I was simply
able to remove the setting of that attribute as it wasn't needed for the
test. In some situations I had to update it to use the correct attribute
for the edition. In some of the tests I was able to remove the entire
factory line as it wasn't needed by the test.

I also removed latest_major_change_published_at as it was never used
execpt in the tests.