Add UI on bulk uploads, and controller response to that UI
We render 2 tabs in the attachments section.  One is the original attachments UI (we call this "individual upload"), the other is a new "Bulk Upload" UI to allow uploading a single zip file.

The tabs are a bit odd in terms of JS, because I want them to be form controls that record the users intent rather than just links.  The controller response is based on which radio button is selected.

If the bulk upload button wasn't selected then we do nothing.  If it is we look at the bulk_upload_zip_file params.  If those params mean the BulkUpload::ZipFile instance is invalid we re-render the form instead of committing the action.  If it's valid, we will proceed to commit the action (there's more to come here to deal with the actual contents of the zip file).

Rationale for this is that when the user has javascript off the form will still work.  A future commit can enhance the bulk upload UI to use AJAX and hit a separate endpoint that only redraws the form.