Turn the bulk upload zip file into params for manipulating attachments
First we remove any edition_attachments_attributes - because choosing to do a bulk upload trumps anything you might have done on the "individual upload" section.

We extract the files from the zip file and look at the attachments on the edition instance in question.  For each file that doesn't match the file name of an existing attachment we turn this into a set of params for adding a new edition_attachment instance.  For each file that does match the name of an existing attachment we turn this into a set of params for replacing the attachment_data of the that attachment.

We then let the action continue on as normal.