Uploading a zip file of bulk attachments forces a review of changes
Added a new method to Attachable (force_review_of_bulk_uploads), calling this gives the model a new attribute (attachments_were_bulk_uploaded).  Setting this attribute to something causes a validation error suggesting that the user review the bulk upload changes.  In BulkUpload::ZipFileToAttachments#manipulate_params! we set a value for this attribute in the params.

We call force_review_of_bulk_uploads on DetailedGuides and Publicationesque edition types.