Factory-built users now have a uid
Without a uid a factory built user cannot be persisted to the session by GDS-SSO.  Our tests would still work because the mock_gds_sso strategy for warden is to always authenticate using the first available user, or whatever is in GDS::SSO.test_user.

We found this because most public facing pages don't actually ask warden to try and log someone in, they just ask if there is already someone there.  If the request to a public page wasn't the first request in the test it would fail because there was nothing in the session.  If it was first it only worked because the Warden::Test::Helpers#login_as helper drops something into warden to say "on the next request, pretend this user is in the session".