Log out to test that old attachments redirect to their replacements
This broke because when we fixed the user factory to allow users to be persisted in the session properly, it meant there was now a user available to see if the attachment belonged to a visible edition.  Anyone that can log in to inside gov is allowed to see any edition of a document (so that preview works) and so, they could see the old attachment because it belonged to an archived edition.  If we logout before trying to see the redirect we get the expected behaviour.

Not saying that this is right, I just wanted the build to be green.  We may need to further think about how we do attachments and urls.