Describe desired bulk upload behaviour
Also implemented the feature fully.  We can choose "Bulk upload" and send a zip file, this then returns the form again forcing us to review the changes made to the attachments by the contents of the zip file.  We review and fix any errors with blank meta-data on new ones and then save the document.