Don't use /tmp for bulk upload zip extraction
Made it configurable via BulkUpload::ZipFile.default_root_directory (much like Whitehall::Uploader::AttachmentCache).  Use an initializer to set it to Rails.root.join('bulk-upload-zip-file-tmp').  We'll use puppet to make this a symlink to a big data store on production.

Configure it during tests to use a different dir and get rake test:cleanup to remove it.