Fill out Parsers::SumariseBody
Turn the body into html using Govspeak, remove the html using ActionView::Base.full_sanitizer.  Then we try to truncate to 140 chars, allowing for the 140th char to be inside a word and doing the truncation at the end.  We've rolled our own because this is the opposite behaviour to rails own truncation.  It is conservative to make sure you don't go over the limit at all whereas our limit is soft so going over it is ok.