Ensure file upload duplication works in IE7
There were two issues here. The first was that the duplication didn't
work as expected due to trying to `unwrap` fields with errors. Changing
the selector on the unwrap now works better.

The second issue was that IE doesn't always fire change events for file
inputs which are dynamically added to the page. This was stopping you
from ever getting more than one file input. I changed the event to a
click which, while causes the new attachment box to generate sooner. Is
more reliable.