Reloading the document series data for decc
This is becuase data was imported with invalid document serieve
missing, instead of haluting the import. So we are reload this data and
add any EditionDocumentSeries which are missing.

NB: I manually had to check and clean up the slugs as the excel method
is not the same as our regex.

Also adding a rails validations to on EditionDocumentSeries