Script to output non document admin -> public urls
I've renamed the existing dump all documents script and extended it to
include many different types of admin links that we need to convert to
public links.

This was given to Robin manually; the right solution will be to build
this into the redirector workflow. At the moment, the
document_mappings.csv is used for two different things: mapping old ->
new urls and mapping admin -> public urls. The data we need in each
overlap but are not the same.

This file provides a union of those as a workaround, but this process
needs revisting soon.