Data migration to set the change note for imported stats
Imported statistical data sets weren't first published at the date of import.  Even if we let them change the date it's not going to say the right thing in the change history popup as it'll say "First published" which may still not be true.  Instead we set the change note on any imported stat to say "Data updated" to avoid this.