Extract info from errors hash and show seperately
We want to ensure that certain 'errors' on edition aren't shown in quite
the same way, so that we can call out particular actions we want the
author to perform. (eg review bulk uploaded attachments).

We are therefore using a ':information' key on the errors has: if this
is set the edition form will now pull it out and show it in an info box.
If this is the only type of error in the errors hash, valid? will still
be false, so the form will be shown, but the 'there are problems with
this document' alert at the top of the screen will not be shown.

Tested implicitly via the feature.

Ticket for this change: https://www.pivotaltracker.com/story/show/42901293

**NOTE:** Whilst working on this we discovered the bulk uploading on
creation of a new edition isn't working. Raised as a seperate bug:

https://www.pivotaltracker.com/story/show/43221595