Extracting params for bulk upload zip file doesn't render on error
Instead of setting a single error on the edition and re-rendering the form when there is an error with the bulk upload zip file, we instead set a param that forces the edition to fail validation.  This means we don't need the @edition instance to pre-exist before we can safely call EdtionsController::Attachments#extract_attachments_from_zip_file.  As per previous commit it will cope with a nil @edition on success and it won't try to use it on failure, leaving it up to params to do the job.