Index additional fields from editions
In order to support filtering we need to add additional fields to the
the elastic search index, such as policies, topics, ministers, id etc.

Rummageable has hard-coded validation of the allowed fields so we have
monkey-patched rummageable to disable the validation. It would be better
to update Rummageable to make it more flexible, or maybe to remove the
validation altogether.

We've added a more thorough unit test for the Searchable mixin.