Adding DocumentFilter::ElaticSearch for publications
This implements an alternative searching and filtering backend for
elastic search.

To use this all the documents need to have been reindexed by rummager
and you need to supply an extra url parameter `use_elastic_search`

There are no tests because this is a prototype. In the production
implementation we intend to use Rummager rather than interfacing
directly with elasticsearch using Tire.