Change method for and testing of govdelivery feeds
Govdelivery feeds were previously toggled in the builder file for each
controller. I have moved that logic and put it into the feed helper.
This means that we no longer need to explicity test govdelivery feeds in
each controller test but can test it once in the feed helper unit tests
and assume it works for each controller. I have however, also left one
test in for it in the `home_controller_test` just to ensure it still
functions as expected on a real request.