Use FakeSearch by default in test
For those tests that want to exercise what they do with "real" search results, we also provide NotQuiteAsFakeSearch.  This supplies a Rummageable implementation to allow storing content and a DocumentFilter implementation to allow fetching content.  This can be invoked at the class level with with_not_quite_as_fake_search.