Make postal addresses more structured
We've introduced separate fields for locality, region. I've renamed
'address' to 'street_address', and 'postcode' to 'postal_code' to match
the vcard spec names. I renamed 'description' to 'title' as this better
reflects its usage to date.

I've added 'country' as this is required to represent international
address which we'll have for the worldwide offices. I've also added a
free-form comments text field to allow input of things such as office
opening hours or instructions on how to contact that office in writing.