Speeches can be imported without delivered on date
If delivered_on is blank, we also throw away any delivered_by information in the file - because we want to make sure we find the correct role for the deliverer based on the date of the speech.  If we don't have the date we can't be sure we have the correct role, so it's better to force the importing user to fix this during speed-tagging.