Use a translations table for world locations
This does everything required to move the name and description of
every WorldLocation into a translations table, into which alternative
translations can be added.  The work to actually add alternative
translations or allow them to be displayed will be added in future
commits.

In general this was straightforward; the main issue was how to sort
locations on values no longer in the main table.  As there the number
of locations is limited, I have resorted to sorting them in ruby but
for other model types, a more efficient solution will need to be
found.