Rename AuditEntry#event to AuditEntry#action
I found it confusing that a PaperTrail Version has an `#event` (create, update
or destroy, by default) and an AuditEntry also had an `#event`. I've renamed
`AuditEntry#event` to `AuditEntry#action` to remove this confusion. I think
that `AuditEntry#action` is also more consistent with our use of
`AuditEntry#actor`.