Fix bug with swapping a lead org to supporting org and changing order
If we move the 1st lead org to be a supporting org, and push the 2nd lead org up to be the 1st we would get errors about duplication.  The problem is that we collect the edition organisation instances from all 3 associations and we weren't reducing the ones we look at for duplication purposes to ignore unchanged instances when we had a changed instance.