Show translated page labels on the world location show view.
We've extracted all of the currently-available page labels into
locale files (initially just English and French to demonstrate the
functionality).

The locale a page is shown in can currently only be set by adding a
`locale=fr` parameter to the URL. How the actual URL is constructed
and the UI for chosing a locale is a separate concern.

It's possible that we'll need to revise the taxonomy of translated
items as other languages with less Latin-compatible pluralisation are
added, but for the moment this is a sufficient first step.

Where any sophisticated construction of translates is required, we
should encapsulate that into helper methods. This will help protect
the view from too much churn as the needs of varying languages come
to bear on the application.

We plan to use Localeapp[1] to administer the translation files, I'll
go into that in more detail in a later commit.

[1]: http://www.localeapp.com