Introduce Localeapp.
The Localeapp gem allows us to push and pull translations to the
[online application][1]. As we extract or adapt the translation
keys, we push up our base (en) translation file to the service, e.g.

    bundle exec localeapp push config/locales/en.yml

On the application, we can then see which translations are missing in
any of the other created locales.

Translators can have accounts on the service, and can be restricted
to editing only specific locales. When they have made translations,
we (the developers) can *pull* their changes down.

    bundle exec localeapp pull

I have made this application private, and we'll need to invite
developers and translators. It appears to be free to use, which is a
bonus.

The changes to the locale files within this commit are a result of
pushing and pulling them from the service, which has added some
translations, and re-ordered the existing ones.

Locale appears to include a YML file in the log directory; it's not
clear if this is an important part of the sync data or if it can be
ignored, so I'm defaulting to ignore it for now.

[1]: http://www.localeapp.com/projects/3437