Replace our use of has_paper_trail
This is the first step toward replacing PaperTrail with the equivalent
hand-rolled functionality, that we believe will play nice with Globalize3.

I've inlined the PaperTrail functionality that we were using into
Edition::AuditTrail. I've left out the `after_destroy` callback as we never
destroy editions and so were never using it. I've also left out the storing of
the serialized object as we were never using that in production code.

*NOTE* Although we weren't using the serialized object in production code, we
were using it in the 'update records the previous version of the document in
the edition version history' test. I've taken the liberty of removing that
test as I don't believe it's necessary.