Remove PaperTrail
We've incorporated just the bits of PaperTrail that we need to satisfy our
requirements for Edition::AuditTrail and so we are now in a position to remove
PaperTrail itself.

This is desirable because our use of PaperTrail doesn't play nicely with
Globalize3, which we're intending to use for i18n.