An alternative translation flow to Localeapp.
It seems like getting disparate, non-technical users to integrate
with Localeapp is a non-starter. That's a shame.

Instead, they will be most effective if they are given spreadsheets
of data to translate.

We can generate this data from the existing translations. This commit
includes a rake task which will export a CSV suitable for spreadsheet
based translation. The command:

    bundle exec rake "translation:export[.,en,fr]"

will export "fr.csv" into the current directory, containing a mapping
of source and translation strings. It should be fairly trivial to
sort the sheet in Excel/Google Docs and see where some translation is
missing.

The translators will still need some training (if only "avoid date and
time translations that you don't understand"), but this could form
the basis of a workflow that is less taxing on the organisation as a
whole.