Make Rummager the default search backend
Pull in the advanced-search-for-whitehall branch of the gds-api-adaptors gem in order to use the new methods.

Notably all the values in the params have to be strings.  They can be arrays or hashes, but the values contained within must be strings.  It might make sense to push this down to gds-api-adaptors.