Remove cache clearing from asset precompilation
Chatting with Edd in our office, the subject of slow asset
precompilation came up.  As we're now considering using a set of
dedicated stylesheets for right-to-left languages, each of which needs
compilation, the current poor performance will only get worse.

James M mentioned this blog post[1] which seems to have a fix.
Running it locally on my machine reduced the time taken for asset
compilation down from 2:21 to 0:26.

[1] http://sj26.com/2013/02/09/the-asset-pipeline-isnt-actually-slow