Allow existing world location translations to be edited
In adding this functionality a couple of changes to the original
translation work were required.  Specifically, I found it difficult
ensuring the location was scoped to the correct locale when getting
and setting attributes.  To get around this, I've introduced a
LocaleWrapper object that binds a model to a given locale, and so
far I'm happy with this approach.

There are still some rough edges in the overall functionality which
I will iron out in the next few commits, specifically:

a) Unable to see the english version alongside the new translation
b) Locations are sorted in memory, not sql on index pages
c) LocaleWrapper needs tidying and extracting