Display links to available translations of an edition
* We are displaying links to all available translations, but it is not
currently possible to create translated editions in the admin interface, so
the links will never show up.

* We are using the "raw" locale name as the link text and not the
human-friendly language name for the translation.

* We are showing links to *all* available translations including the current
translation. We plan to "disable" the link to the current translation in a
future commit.

* We've had to add an es.yml file, containing the format of the long_ordinal
date, to avoid our viewing-published-international-priorities feature from
exploding with the following error:

    translation missing: es.date.formats.long_ordinal (ActionView::Template::Error)