Avoid serving English editions on multiple URLs
I can imagine that we might want something like this across the site
eventually, i.e. any request that specifies 'locale=en' should probably
redirect to the same URL without the locale specified. This is a good enough
starting point for now, though.

Having written the above, it's possible that we'll change the URL structure
such that it always contains the locale, in which case this problem goes away.

*NOTE* I've had to include PublicDocumentRoutesHelper in DocumentsController
so that I can use `public_document_path`.