Translate labels shown on priority pages.
This will also have the effect of translating labels on many other
edition type pages.

The translation keys are evolving slightly here; I'll likely rework
the document types to include pluralisation in a similar style.

I didn't add a `key` attribute to OrganisationType, but we could
consider doing this in the future, if only for consistency of
translation keys.

I had to modify some existing tests because the factory-auto-created
`OrganisationType` instances cannot be sorted in listing order. It's
actually a bit unrealistic to generate new `OrganisationType` instances,
since the set in the DB is not editable. They are basically constants,
and are probably a candidiate for the 'ActiveRecord-esque'
implementation used by `PublicationType` etc.