Avoid use of capitalisation in translations.
Also, move towards a consistent way of translating types that will
support pluralisation.

It's worth noting that to some extent I've just swapped a call to
`capitalize` with a call to `downcase`; however, I think this is the
right direction to go in, and we can also push out the call to
`downcase` by translating the full 'see all' strings later.