Add the ability for writers to translate editions
We've added the basic functionality to allow writers and editors to create
translations of non-published documents. The translation is limited to the
title, summary and body at present.

We've taken a lot of inspiration from @tomafro's work on adding translations
to world locations.

We've followed the pattern of allowing document types to specify which
behaviour they have and introduced the `Edition.translatable?` method. We're
currently only allowing International Priorities to be translated.

We've chosen to show the available translations in a similar fashion to
supporting pages, i.e. in their own tab and presented as a 'page' per
translation. *NOTE* The ability to edit translations will come later.