Set database default character set and collation
It appears that the default character set for the production database
is currently latin1, not utf8.  We need to store data in utf8 so
we can accommodate an extended range of languages (including urdu,
arabic, japanese, farsi, etc...).  Explicitly setting the defaults
ensures that new tables will be created with a utf8 character set.

Explicitly setting the migration timestamp to the first of Jan ensures
that the migration will run before any other outstanding migrations
currently outstanding on production.  This is important, as these
later migrations may otherwise create latin1 tables that the
convert-tables.sh script is not set up to fix.