Remove JavaScript relative times
The main reasons for removal:

- Relative times posed a problem with translations. We would have had to
  translate the javascript for different locales, which why wasn't
  impossible was an extra pain.
- Relative times were really unhelpful when it gets beyond 1 or 2 days
  ago. If you are searching for something old you really what to know
  when it was not a rough 'x months ago'
- Publications, consultations and speechs all had incorrect times.
  Those format only have dates for their first edition, so the time
  ago was calculated from midnight on the day. This has caused
  confusion and bugs being reported in the past.

Removed the `relative_time` helper which was used to render absolute
times and convert them to relative times on the in JavaScript.

Converted the uses of `relative_time` to `absolute_time` or
`absolute_date` where it made sense.

Removed the calls to `timeago` from `application.js`. It would seem it
was never called in the admin so switching to absolute times will have
no visual difference.