Make edition translations migration deployment-friendly.
* For a while after deployment there will be application processes running
using the old codebase. Removing the columns on the editions table would mean
that errors would occur when the old code tried to access the title, summary
and body attributes via the editions table.

* We have *edited* the original migration so that it does *not* remove the
columns on the editions table. However, since this original migration has
already run on CI, preview and most development machines, we've also added a
migration which conditionally adds the columns back onto the editions table if
they have already been removed i.e. by the original version of the original
migration. In this way, we should end up with all machines having the columns
on both editions and edition_translation tables and CI & preview should
provide meaningful results to inform a deployment to production.

* We've committed the schema.rb as it appears after running the migrations
before editing them in this commit. The schema on production will be different
in that the title, summary and body columns on the editions table will be a
different order. We're not too worried about this as those columns are going
to be removed in the near future.

* We have also checked that Globalize3 works ok even if the title, summary and
body columns are present on the editions table.